// frontend/src/app/models/player.ts
export interface PlayerModel {
    id: number;
    fifa_version: string;
    fifa_update: string;
    player_face_url: string;
    long_name: string;
    club_name: string;
    player_positions: string;
    nationality_name: string;

    overall: number;
    potential: number;

    value_eur: number;
    wage_eur: number;

    age: number;
    height_cm: number;
    weight_kg: number;

    preferred_foot: string;
    weak_foot: number;
    skill_moves: number;
    international_reputation: number;

    work_rate: string;
    body_type: string;

    // Core
    pace: number;
    shooting: number;
    passing: number;
    dribbling: number;
    defending: number;
    physic: number;

    // Attacking
    attacking_crossing: number;
    attacking_finishing: number;
    attacking_heading_accuracy: number;
    attacking_short_passing: number;
    attacking_volleys: number;

    // Technical
    skill_dribbling: number;
    skill_curve: number;
    skill_fk_accuracy: number;
    skill_long_passing: number;
    skill_ball_control: number;

    // Movement
    movement_acceleration: number;
    movement_sprint_speed: number;
    movement_agility: number;
    movement_reactions: number;
    movement_balance: number;

    // Power
    power_shot_power: number;
    power_jumping: number;
    power_stamina: number;
    power_strength: number;
    power_long_shots: number;

    // Mentality
    mentality_aggression: number;
    mentality_interceptions: number;
    mentality_positioning: number;
    mentality_vision: number;
    mentality_penalties: number;
    mentality_composure: number;

    // Defending
    defending_marking: number;
    defending_standing_tackle: number;
    defending_sliding_tackle: number;

    // Goalkeeping
    goalkeeping_diving: number;
    goalkeeping_handling: number;
    goalkeeping_kicking: number;
    goalkeeping_positioning: number;
    goalkeeping_reflexes: number;
    goalkeeping_speed: number;

    player_traits: string;
}